fix: return dates as string instead of date object
Without these config values, the dates are automatically parsed and returned as date objects. Even though it could be useful in certain cases, it is a problem for our use case because it converts the date to another timezone (it removes 2 hours, so it must convert from local to UTC or something).
By returning dates as strings, we avoid this conversion, and we can later manually make a date object with the string (`new Date(<our string date>)`) which will have no timezone issue.