feat: use dynamic camera port to fetch resources from correct camera
To know what camera the resources have to be fetched from, we require the the camera id to be passed as parameter in the related API urls.

Then, we use the `requireCameraMiddleware` middleware function to ensure we get a valid camera id, and if so, to get the camera data.
We then pass the camera data to the handler function for this route, which can then use the camera port to fetch the resources on the right camera.