fix: array must not be enclosed in [] if passed in paranthesis in query
In the links I provided, it is said to enclose arrays in brackets when passing them to the query parameters for them to work.

I don't know if it's true or not, but enclosing the array in brackets while having the "WHERE ... IN (?)" statement did not work and produced the following error:

`DELETE FROM notification_subscription  WHERE FK_user_id = 1 AND subscription_id IN ((42, 43))` - Operand should contain 1 column(s)

It seemed to put two pair or parenthesis, which does not work.
It looks like the solution can either be:
1. use the statement "WHERE ... IN ?" without parenthesis, and enclose the array in brackets when passing it as query parameter
2. use the statement "WHERE ... IN (?)" with parenthesis, and not enclose the array in brackets when passing it as query parameter