feat: update validation logic of `isObjectEmpty()`
For our validation code to work, we have first to make sure  that the parameter we receive is an object.

A lot of different values are of type "object", such as Map(), Set(), null, etc.

We account for that by strictly checking the parameter against a null value, then by ensuring the object has a `.hasOwnProperty()` method.
If not, the returned value of the check will equals to `undefined` and the function will ultimately return `true`.