feat: return non-formatted data as better practice
It is a good practice to return the data, almost raw, without any specific format or organization, because the use cases we will use this data for can be different, and one use case might not want a predefined format that another use case needs.

In our case, we do this change specifically because before we were relying on the detection_session_id being always in the same order as the recording time, thus returning data correctly sorted.
But this is no more true with the changes made in the camera project, where detection_sessions can be created in the database a long time after the recording was already made.